feat: Integrate React Router and refactor components into a dedicated HomePage
