+++ package.json
@@ -14,7 +14,8 @@
     "bootstrap": "^5.3.8",
     "react": "^19.1.1",
     "react-bootstrap": "^2.10.10",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "react-router": "^7.9.3"
   },
   "devDependencies": {
     "@eslint/js": "^9.36.0",
